pyproject.toml
[project]
name = "pre-processing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fiona>=1.10.1",
    "geopandas>=1.1.1",
    "loguru>=0.7.3",
    "pandas>=2.3.3",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
]
